feat: add reusable glass tile widget
